pyproject.toml
[project]
name = "paper-check"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "langchain-google-genai>=2.1.10",
    "langgraph>=0.6.7",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.6",
    "python-multipart>=0.0.9",
]
